requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
abseil-cpp=20230802.0=h313beb8_2
absl-py=2.1.0=py310hca03da5_0
addict=2.4.0=pyhd8ed1ab_2
affine=2.3.0=pyhd3eb1b0_0
antlr-python-runtime=4.9.3=pyhd8ed1ab_1
attrs=24.3.0=py310hca03da5_0
blas=2.131=openblas
blas-devel=3.9.0=31_h11c0a38_openblas
blosc=1.21.3=h313beb8_0
boost-cpp=1.78.0=hf1d6563_2
bottleneck=1.4.2=py310hbda83bc_0
brotli-python=1.0.9=py310h313beb8_9
bzip2=1.0.8=h80987f9_6
c-ares=1.34.4=h5505292_0
ca-certificates=2025.2.25=hca03da5_0
cairo=1.16.0=h302bd0f_5
certifi=2025.1.31=py310hca03da5_0
cffi=1.17.1=py310h3eb5a62_1
cfitsio=4.1.0=hd4f5c17_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py310hca03da5_0
click-plugins=1.1.1=pyhd3eb1b0_0
cligj=0.7.2=pyhd3eb1b0_0
contourpy=1.3.1=py310h48ca7d4_0
cpython=3.10.16=py310hd8ed1ab_2
curl=8.12.1=h3e2b118_0
cycler=0.11.0=pyhd3eb1b0_0
expat=2.7.0=h286801f_0
filelock=3.13.1=py310hca03da5_0
fiona=1.8.22=py310hcdcf461_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.1=hee714a5_2
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.55.3=py310h80987f9_0
freetype=2.13.3=h47d26ad_0
freexl=1.0.6=h1a28f6b_0
fsspec=2024.12.0=py310hca03da5_0
future=1.0.0=py310hca03da5_0
gdal=3.5.2=py310h0b34360_4
geos=3.11.0=h9a09cb3_0
geotiff=1.7.1=hc898e3f_3
gettext=0.23.1=h3dcc1bd_0
gettext-tools=0.23.1=h493aca8_0
giflib=5.2.2=h80987f9_0
glib=2.78.1=h9e231a4_0
glib-tools=2.78.1=h9e231a4_0
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py310h5c1b81f_0
grpc-cpp=1.48.2=hc60591f_4
grpcio=1.48.2=py310hc60591f_4
gtest=1.14.0=h48ca7d4_1
hdf4=4.2.15=h1a38d6a_5
hdf5=1.12.2=nompi_ha7af310_101
huggingface_hub=0.29.2=py310hca03da5_0
hydra-core=1.3.1=pyhd8ed1ab_0
icu=70.1=h6b3803e_0
idna=3.7=py310hca03da5_0
imagecodecs=2025.3.30=pypi_0
imageio=2.37.0=pypi_0
importlib_resources=6.5.2=pyhd8ed1ab_0
jinja2=3.1.6=py310hca03da5_0
jpeg=9e=h80987f9_3
json-c=0.16=h1a28f6b_0
kealib=1.4.15=h02ce806_1
kiwisolver=1.4.8=py310h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lazy-loader=0.4=pypi_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libaec=1.1.3=h313beb8_0
libasprintf=0.23.1=h493aca8_0
libasprintf-devel=0.23.1=h493aca8_0
libblas=3.9.0=31_h10e41b3_openblas
libcblas=3.9.0=31_hb3479ef_openblas
libcurl=8.12.1=hde089ae_0
libcxx=20.1.2=ha82da77_0
libdap4=3.20.6=h8510809_2
libdeflate=1.8=h1a28f6b_5
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libexpat=2.7.0=h286801f_0
libffi=3.4.4=hca03da5_1
libgdal=3.5.2=h1ecf2db_4
libgettextpo=0.23.1=h493aca8_0
libgettextpo-devel=0.23.1=h493aca8_0
libgfortran=5.0.0=14_2_0_h6c33f7e_2
libgfortran5=14.2.0=h6c33f7e_2
libglib=2.78.1=hd9b11f9_0
libiconv=1.18=hfe07756_1
libintl=0.23.1=h493aca8_0
libintl-devel=0.23.1=h493aca8_0
libjpeg-turbo=2.0.0=h1a28f6b_0
libkml=1.3.0=hc4d7c42_7
liblapack=3.9.0=31_hc9a63f6_openblas
liblapacke=3.9.0=31_hbb7bcf8_openblas
liblzma=5.8.1=h39f12f2_0
libnetcdf=4.8.1=nompi_h2510be2_106
libnghttp2=1.57.0=h62f6fdd_0
libopenblas=0.3.29=openmp_hf332438_0
libpng=1.6.39=h80987f9_0
libpq=14.5=h1a28acd_5
libprotobuf=3.20.3=h514c7bf_0
librttopo=1.1.0=h275bb25_11
libspatialite=5.0.1=h97b496e_19
libsqlite=3.46.0=hfb93653_0
libssh2=1.11.1=h3e2b118_0
libtiff=4.4.0=h2fd578a_2
libuv=1.48.0=h80987f9_0
libwebp-base=1.3.2=h80987f9_1
libxcb=1.15=hf346824_0
libxml2=2.10.3=h67585b2_4
libzip=1.10.1=ha0bc3c6_3
libzlib=1.2.13=hfb2fe0b_6
lightning-utilities=0.11.9=py310hca03da5_0
llvm-openmp=20.1.2=hdb05f8b_0
llvmlite=0.44.0=pypi_0
lz4-c=1.9.4=h313beb8_1
markdown=3.4.1=py310hca03da5_0
markupsafe=3.0.2=py310h80987f9_0
matplotlib=3.6.2=py310hca03da5_0
matplotlib-base=3.6.2=py310h8bbb115_0
minizip=4.0.3=ha89c15f_0
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py310hca03da5_0
munch=4.0.0=pyhd8ed1ab_1
ncurses=6.5=h5e97a16_3
networkx=3.3=py310hca03da5_0
ninja=1.12.1=hca03da5_0
ninja-base=1.12.1=h48ca7d4_0
nspr=4.35=h313beb8_0
nss=3.89.1=h313beb8_0
numba=0.61.2=pypi_0
numexpr=2.10.1=py310h5d9532f_0
numpy=1.26.4=py310h3b2db8e_0
numpy-base=1.26.4=py310ha9811e2_0
omegaconf=2.3.0=pyhd8ed1ab_0
openblas=0.3.29=openmp_hea878ba_0
opencv-python=4.11.0.86=pypi_0
openjpeg=2.5.0=h5d4e404_1
openssl=3.1.8=h81ee809_0
optree=0.14.1=py310h7f4e7e6_1
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
pcre=8.45=hc377ac9_0
pcre2=10.40=hb34f9b4_0
pillow=11.0.0=py310hfaf4e14_0
pip=25.0=py310hca03da5_0
pixman=0.44.2=h2f9eb0b_0
poppler=22.10.0=hae7f5f0_0
poppler-data=0.4.11=hca03da5_1
postgresql=14.5=h45c140d_5
proj=9.0.1=h4c79c2b_1
protobuf=3.20.3=py310h313beb8_0
pthread-stubs=0.3=h1a28f6b_1
pybind11=2.13.6=pyh1ec8472_2
pybind11-global=2.13.6=pyh415d2e4_2
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pyhd8ed1ab_1
pyparsing=3.2.0=py310hca03da5_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.13=h2469fbe_0_cpython
python-dateutil=2.9.0post0=py310hca03da5_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.10=6_cp310
pytorch=2.1.0=gpu_mps_py310h87e4ab7_100
pytorch-lightning=1.7.0=pyhd8ed1ab_0
pytz=2024.1=py310hca03da5_0
pyyaml=6.0.2=py310hc74094e_2
rasterio=1.3.2=py310ha36aacf_0
re2=2022.04.01=hc377ac9_0
readline=8.2=h1a28f6b_0
requests=2.32.3=py310hca03da5_1
safetensors=0.5.3=py310hf4b39b5_0
scikit-image=0.25.2=pypi_0
scipy=1.15.2=py310h14fe6a6_1
setuptools=58.0.4=py310hca03da5_0
shapely=1.8.5=py310hc253cc3_1
six=1.17.0=pyhd8ed1ab_0
sleef=3.5.1=h80987f9_2
snappy=1.2.1=h313beb8_0
sngp-pytorch=0.0.1.dev13=pypi_0
snuggs=1.4.7=pyhd3eb1b0_0
sqlite=3.45.3=h80987f9_0
sympy=1.13.3=pyh2585a3b_105
tensorboard=2.19.0=py310hca03da5_0
tensorboard-data-server=0.7.0=py310ha6e5c4f_1
tifffile=2024.12.12=py310hca03da5_0
tiledb=2.11.3=h9bd36d0_1
timm=1.0.15=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
torchmetrics=0.7.0=pyhd8ed1ab_0
torchvision=0.15.2=cpu_py310h31aa045_0
tornado=6.4.2=py310h80987f9_0
tqdm=4.67.1=py310h33ce5c2_0
typing-extensions=4.13.0=h9fa5a19_1
typing_extensions=4.13.0=pyh29332c3_1
tzcode=2025b=h5505292_0
tzdata=2025a=h04d1e81_0
unicodedata2=15.1.0=py310h80987f9_1
uriparser=0.9.8=h00cdb27_0
urllib3=2.3.0=py310hca03da5_0
wand=0.6.13=pypi_0
werkzeug=3.1.3=py310hca03da5_0
wheel=0.45.1=py310hca03da5_0
xerces-c=3.2.4=h627aa08_1
xorg-libxau=1.0.12=h5505292_0
xorg-libxdmcp=1.1.5=hd74edd7_0
xz=5.6.4=h80987f9_1
yaml=0.2.5=h3422bc3_2
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=hfb2fe0b_6
zstd=1.5.6=hfb09047_0
